requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.5
django-choices==1.7.2
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
jwcrypto==1.3.1
kombu==5.2.4
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.1
redis==4.3.3
requests==2.28.0
requests-oauthlib==1.3.1
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tweepy==4.10.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
